Crate bevy_hierarchy

bevy_hierarchy can be used to define hierarchies of entities.

Most commonly, these hierarchies are used for inheriting Transform values from the Parent to its Children.

Structs

AddChild

Command that adds a child to an entity

ChildBuilder

Struct for building children onto an entity

Children

Contains references to the child entities of this entity

DespawnChildrenRecursive

Despawns the given entity’s children recursively

DespawnRecursive

Despawns the given entity and all its children recursively

HierarchyPlugin

The base plugin for handling Parent and Children components

InsertChildren

Command that inserts a child at a given index of a parent’s children, shifting following children back

Parent

Holds a reference to the parent entity of this entity. This component should only be present on entities that actually have a parent entity.

PushChildren

Command that pushes children to the end of the entity’s children

RemoveChildren

Command that removes children from an entity, and removes that child’s parent and inserts it into the previous parent component

WorldChildBuilder

Struct for adding children to an entity directly through the World for use in exclusive systems

Enums

HierarchyEvent

A Event that is fired whenever there is a change in the world’s hierarchy.

Traits

BuildChildren

Trait defining how to build children

BuildWorldChildren

Trait that defines adding children to an entity directly through the World

DespawnRecursiveExt

Trait that holds functions for despawning recursively down the transform hierarchy

Functions

despawn_with_children_recursive

Function for despawning an entity and all its children